dateTime conversion in the excelimporter
This patch brings proper conversion for values passed as date time
from .xls to .ods. The proper endering in KSpread is not there yet,
though.

svn path=/trunk/koffice/; revision=1063020

+++ kspread/Value.cpp
@@ -476,7 +476,7 @@ QDateTime Value::asDateTime(const CalculationSettings* settings) const
     QDateTime datetime(settings->referenceDate(), QTime(), Qt::UTC);
 
     const int days = asInteger();
-    const int msecs = qRound((numToDouble(asFloat() - days)) * 86400000.0);      // 24*60*60*1000
+    const int msecs = qRound((numToDouble(asFloat() - double(days))) * 86400000.0);      // 24*60*60*1000
     datetime = datetime.addDays(days);
     datetime = datetime.addMSecs(msecs);
